Extract CP864 to its own file
